Add basic database functionality
This will be used to save the pomodoro history.

startup() will attempt to create a "pomodoros" table if it doesn't
exist, so that if the database is already there there's no risk of the
table being overwritten.

shutdown() makes sure that the changes are committed to the database and
the connection is closed.