Complete do_pomodoro() function
The function gets the name the task will be saved as from user input,
and from then the program runs without further input. There's a 5 second
wait in which the pomodoro can be cancelled without it counting as a
cancelled pomodoro - I may want to make it obvious when that wait has
ended.

I still haven't sorted a ding sound - easy enough to make a beep on
Windows but my motherboard has no buzzer so I can't just use an alert
character. I'll probably end up calling an external program, and that
can be tweaked for a Windows version.

Ran into a problem with sqlite when attempting to store the length
timedelta. sqlite does not support that data type, so I had to convert
it into a string - not ideal but it should be easy enough to convert
back when analysing. I bet the datetime module has something for doing
that.

Inserting into the database seems to now be fully functional.