Fix sound playback on windows
At least on my windows install, sox was not auto-detecting the default
sound device and it had to be explicitly specified. In the README and as a
comment in the code, I've mentioned that the specified device may need to
be changed - it could be outputting on headphones rather than speakers,
etc.