Add fundamental command line functionality
Implemented:
    Normal pomodoro creation
    Repeat pomodoro creation with -r
    Pomodoro listing with -l (last n pomodoros, default 5)
    Pomodoro export with -e
    Pomodoro length setting with -t

All this functionality can be accessed from the main pomodoro.py,
although some of it (like exporting) can be used by running the specific
file.

This is now a fully functional replacement for the version I was
previously using - all functionality that existed works with the new
command-line interface.