Add pomodoro name to pomodoro start message
When using the "-r" option, I noticed I always did a "-l" beforehand
just to make sure that it was what I thought it was. Rather than doing
this, it seems to make sense to just say the name of the pomodoro along
with the 5 second warning, since it was already being passed to the
function.