Add basic structure of analyse_pomodoros()
This is the function that will find and print statistics about the
pomodoros passed to it. As of right now there are three different
functionalities it will be able to do: breaks, frequency and complete.

"breaks" will analyse the time in between pomodoros. It'll ignore gaps
over a certain time (2 hours?) and find an average, and maybe rolling
averages.

"frequency" will analyse how often the pomodoros are done. Depending on
the time span of the pomodoros it could group by day, week, month or
year. It could also show average statistics per day/week/month. I'll
likely need to import a terminal graphing library.

"complete" will analyse how many complete and incomplete pomodoros there
are. It will be able to give statistics for things like the shortest and
longest incomplete pomodoros, or maybe the average incomplete pomodoro
length.

Additional minor formatting changes.