Add break analysis functionality
What this has to do is find the average time spent between pomodoros. It
can basically throw away anything which is greater than an hour, since
that's quite a long time to call a "break between pomodoros".

The parts of analyse_pomodoros() can add onto the added output string,
so that it can be saved to a file if so desired