Add count functionality to analyse_pomodoros()
The first thing this required was a filter_topic() function to get only
the pomodoros beginning with a certain string ("topic"). This'll
probably be re-used in my filter functionality when it gets written - at
that point it might be moved into a pomodoro_export.py. On second
thought, I end up with one calling it from the other either way so it
may as well stay where it is.

I don't like how I'm specifying the topic right now - it requires using
the name parameter, which has to go before "-a count". I'm not sure what
to do about that right now - I'll have to look at the argparse
documentation some more.