Add "topic" analysis functionality
As the name suggests, this breaks down all pomodoros by topic. It uses
the same function used in count - it might be a good idea to merge the
two. If a name is specified it could count that topic, and otherwise
count all topics.

Fixed a minor bug - filter_topic() didn't check for a space after the
topic, so a pomodoro name beginning with "Steve's" would also be
included when filtered by "Steve".